pyproject.toml
[tool.poetry]
name = "somatic"
version = "0.1.0"
description = "A Python CLI tool that watches Postgres and automatically generates embeddings"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "somatic"}]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
psycopg2-binary = "^2.9.9"
openai = "^1.12.0"
qdrant-client = "^1.7.3"
pydantic = "^2.6.1"
pydantic-settings = "^2.1.0"
rich = "^13.7.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
somatic = "somatic.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
